+++ src/contexts/prefix.jl
@@ -96,6 +96,67 @@ function prefix_and_strip_contexts(ctx::AbstractParentContext, vn::VarName)
     return vn, setchildcontext(ctx, new_ctx)
 end
 
+"""
+    DynamicPPL.extract_prefixes(ctx::AbstractContext)
+
+Recursively process the context stack `ctx` to remove all levels of `PrefixContext`, and
+aggregate their prefixes into a single VarName which can be used to apply all the prefixes
+at once.
+
+Returns a tuple of:
+
+- `ctx_without_prefixes`: the context stack with all `PrefixContext`s removed
+- `vn`: the VarName that contains all the aggregated prefixes that were removed, or
+  `nothing` if there were no `PrefixContext`s in the stack.
+
+# Example
+
+```jldoctest
+julia> using DynamicPPL: extract_prefixes, PrefixContext, AbstractContext, @varname
+
+julia> struct MyContext <: AbstractContext end
+
+julia> ctx = PrefixContext(@varname(b), PrefixContext(@varname(a), MyContext()));
+
+julia> extract_prefixes(ctx)
+(MyContext(), b.a)
+```
+
+!!! warning
+
+    This does _not_ modify any variables in any `CondFixContext`s that may be present in the
+    context stack. This is because this function is only used in functions such as
+     `tilde_assume!!`, which are lower in the tilde-pipeline than `contextual_isassumption`
+    and `contextual_isfixed` (the functions which actually use the `CondFixContext`'s
+    values). Thus, by this time, any `CondFixContext`s present have already served their
+    purpose.
+
+    If you call this function, you must therefore be careful to ensure that you _do not_
+    need to modify any inner `CondFixContext`s. If you _do_ need to modify them, then you
+    may need to use `prefix_cond_and_fixed_variables` instead.
+"""
+function extract_prefixes(ctx::AbstractContext)
+    new_ctx, new_optic = _extract_prefixes(ctx, AbstractPPL.Iden())
+    return if new_optic === AbstractPPL.Iden()
+        new_ctx, nothing
+    else
+        new_ctx, AbstractPPL.optic_to_varname(new_optic)
+    end
+end
+function _extract_prefixes(ctx::PrefixContext, optic::AbstractPPL.AbstractOptic)
+    new_child_context, child_optic = _extract_prefixes(childcontext(ctx), optic)
+    this_optic = AbstractPPL.varname_to_optic(ctx.vn_prefix)
+    return (new_child_context, Base.cat(this_optic, child_optic))
+end
+function _extract_prefixes(ctx::AbstractParentContext, optic::AbstractPPL.AbstractOptic)
+    # ctx is not a PrefixContext, so just need to recurse on the child context
+    new_child_context, optic = _extract_prefixes(childcontext(ctx), optic)
+    return setchildcontext(ctx, new_child_context), optic
+end
+function _extract_prefixes(ctx::AbstractContext, optic::AbstractPPL.AbstractOptic)
+    return ctx, optic
+end
+
 function tilde_assume!!(
     context::PrefixContext,
     right::Distribution,
@@ -111,15 +172,17 @@ function tilde_assume!!(
     # `PrefixContext{:a}(PrefixContext{:b}(DefaultContext()))`, then the
     # first call would apply the prefix `a.b._`, and the recursive call
     # would apply the prefix `b._`, resulting in `b.a.b._`.
-    # This is why we need a special function, `prefix_and_strip_contexts`.
-    new_vn, new_context = prefix_and_strip_contexts(context, vn)
-    # Add 1 for the top-level symbol in the VarName.
-    # NOTE(penelopeysm): I tried to move this into an inner constructor of PrefixContext, so
-    # that it could be reused here and in store_coloneq_value!!, and also just because it
-    # makes sense to tie this information to the PrefixContext. But that caused nonzero
-    # allocations on the LogDensityFunction submodel test, for reasons that are rather
-    # unclear! Be careful if you think of doing that.
-    n = optic_skip_length(AbstractPPL.getoptic(context.vn_prefix)) + 1
+    # This is why we need a special function, `extract_prefixes`.
+    new_context, aggregated_prefix_vn = extract_prefixes(context)
+    new_vn = AbstractPPL.prefix(vn, aggregated_prefix_vn)
+    # Figure out how many levels of templating to skip for this variable (and add 1 for the
+    # top-level symbol in the VarName).
+    # NOTE(penelopeysm): I tried to move this computation into an inner constructor of
+    # PrefixContext, so that it could be reused here and in store_coloneq_value!!, and also
+    # just because it makes sense to tie this information to the PrefixContext. But that
+    # caused nonzero allocations on the LogDensityFunction submodel test, for reasons that
+    # are rather unclear! Be careful if you think of doing that.
+    n = optic_skip_length(AbstractPPL.getoptic(aggregated_prefix_vn)) + 1
     return tilde_assume!!(new_context, right, new_vn, SkipTemplate{n}(template), vi)
 end
 
@@ -134,19 +197,24 @@ function tilde_observe!!(
     # In the observe case, unlike assume, `vn` may be `nothing` if the LHS is a literal
     # value. For the need for prefix_and_strip_contexts rather than just prefix, see the
     # comment in `tilde_assume!!`.
-    new_vn, new_context = if vn !== nothing
-        prefix_and_strip_contexts(context, vn)
+    new_vn, new_context, new_template = if vn !== nothing
+        new_context, aggregated_prefix_vn = extract_prefixes(context)
+        n = optic_skip_length(AbstractPPL.getoptic(aggregated_prefix_vn)) + 1
+        prefixed_vn = AbstractPPL.prefix(vn, aggregated_prefix_vn)
+        prefixed_vn, new_context, SkipTemplate{n}(template)
     else
-        vn, childcontext(context)
+        # We don't need a template at all in this branch, so may as well avoid trying to
+        # choose one.
+        vn, childcontext(context), NoTemplate()
     end
-    n = optic_skip_length(AbstractPPL.getoptic(context.vn_prefix)) + 1
-    return tilde_observe!!(new_context, right, left, new_vn, SkipTemplate{n}(template), vi)
+    return tilde_observe!!(new_context, right, left, new_vn, new_template, vi)
 end
 
 function store_coloneq_value!!(
     context::PrefixContext, vn::VarName, right::Any, template::Any, vi::AbstractVarInfo
 )
-    new_vn, new_context = prefix_and_strip_contexts(context, vn)
-    n = optic_skip_length(AbstractPPL.getoptic(context.vn_prefix)) + 1
+    new_context, aggregated_prefix_vn = extract_prefixes(context)
+    new_vn = AbstractPPL.prefix(vn, aggregated_prefix_vn)
+    n = optic_skip_length(AbstractPPL.getoptic(aggregated_prefix_vn)) + 1
     return store_coloneq_value!!(new_context, new_vn, right, SkipTemplate{n}(template), vi)
 end